+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### General changes:
 * Fixed that players drop too less items with partial inventory loss under rare circumstances
 * Added a configurable "main inventory loss" to the partial inventory loss
+* Transferring items via shift-clicking in the player's inventory during PvP is now disableable
 
 ### SuffixForge compatibility:
 * Added a configuration option to prevent soulbound items to be dropped with the partial inventory loss